requirement.txt
djangorestframework
markdown
django-filter
django-cors-headers
pillow
